requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-extra-fields==2.0.5
django-filter==2.3.0
django-geojson==3.0.0
django-geojson-serializer==0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-gis==0.15
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
